+++ src/screens/hardware-wallets/components/Layout.tsx
