+++ src/screens/hardware-wallets/components/NanoXDeviceAnimation.tsx
